+++ nodes/NooviChat/descriptions/CannedResponseDescription.ts
@@ -127,7 +127,7 @@ export const CannedResponseFields: INodeProperties[] = [
 			},
 		},
 		default: false,
-		description: 'Retornar todos os resultados ou limite',
+		description: 'Whether to return all results instead of applying a limit',
 	},
 	{
 		displayName: 'Limit',